+++ src/iw_val_store.c
@@ -96,7 +96,7 @@ void iw_val_destroy(iw_val *value) {
 
 // --------------------------------------------------------------------------
 
-bool iw_val_to_str(iw_val *value, char *buff, int buff_len) {
+bool iw_val_to_str(iw_val *value, char *buff, size_t buff_len) {
     switch(value->type) {
     case IW_VAL_TYPE_NUMBER :
         snprintf(buff, buff_len, "%d", value->v.number);
